SUMMARY:   Fixing compile bug on unity (tlcc)
     BUGS:      n/a

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  no

     COMMENTS:

     The pgi compiler wanted a cast of an integer to double in fabs().
     This is in a minor print routine, so I didn't run the tests.

(checklist version 2007-06-20)

+++ packages/ml/src/Utils/ml_utils.c
@@ -2511,7 +2511,7 @@ void ML_print_align(int int2match, char *space, int pad)
   int jj;
   double dtemp;
   if (int2match < 0) pad++;
-  dtemp=fabs(int2match);
+  dtemp=fabs((double)int2match);
   for (jj=0; jj<pad; jj++) space[jj]=' ';
   jj=pad;
   while (dtemp>=1) {space[jj++]=' '; dtemp /= 10;}
